pyproject.toml
[tool.poetry]
name = "themagnificent7"
version = "0.1.0"
description = "API endpoint to return the 7-a-side team with the highest magnificence."
authors = ["Harinder"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
djangorestframework = "^3.15.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
